docs(PPDSC-2703): move form complete
