+++ src/form/__tests__/form-input.stories.tsx
@@ -23,11 +23,7 @@ import {
   TextBlock,
   ThemeProvider,
 } from '../..';
-import {
-  IconFilledAddCircle,
-  IconFilledRemoveRedEye,
-  IconFilledStop,
-} from '../../icons';
+import {IconFilledAddCircle} from '../../icons';
 import {SelectOption} from '../../select';
 import {Fieldset} from '../../fieldset';
 import {RadioGroup} from '../../radio-button';
@@ -478,247 +474,6 @@ export const StoryFormInputCharacterCount = () => (
   </Container>
 );
 StoryFormInputCharacterCount.storyName = 'Form input character count';
-
-const FormContainer = styled.div`
-  height: 700px;
-  overflow: auto;
-  max-width: 500px;
-  padding: 20px;
-  margin: 0 auto;
-  border: 1px solid #ccc;
-  border-radius: 5px;
-`;
-
-const ShowPasswordButton = ({
-  onClick,
-  isVisible,
-}: {
-  onClick: () => void;
-  isVisible: boolean;
-}) => (
-  <IconButton
-    aria-label="toggle password"
-    onClick={onClick}
-    size="small"
-    overrides={{stylePreset: 'iconButtonMinimalPrimary'}}
-  >
-    {isVisible ? (
-      <IconFilledStop overrides={{size: 'iconSize010'}} />
-    ) : (
-      <IconFilledRemoveRedEye overrides={{size: 'iconSize010'}} />
-    )}
-  </IconButton>
-);
-
-export const StoryFormComplete = () => {
-  const [showPassword, toggleShowPassword] = React.useState(false);
-  return (
-    <FormContainer>
-      <Form onSubmit={console.log}>
-        <Fieldset legend="Personal">
-          <Block marginBlockStart="space020" marginBlockEnd="space050">
-            <FormInput
-              name="first-name"
-              rules={{
-                required: 'Required field',
-                minLength: {
-                  value: 2,
-                  message: 'Name must be at least 2 characters long',
-                },
-              }}
-            >
-              <FormInputLabel>First name</FormInputLabel>
-              <FormInputTextField />
-              <FormInputAssistiveText>
-                Enter your first name
-              </FormInputAssistiveText>
-            </FormInput>
-          </Block>
-
-          <Block marginBlockStart="space020" marginBlockEnd="space050">
-            <FormInput
-              name="last-name"
-              rules={{
-                required: 'Required field',
-                minLength: {
-                  value: 2,
-                  message: 'Name must be at least 2 characters long',
-                },
-              }}
-            >
-              <FormInputLabel>Last name</FormInputLabel>
-              <FormInputTextField />
-              <FormInputAssistiveText>
-                Enter your last name
-              </FormInputAssistiveText>
-            </FormInput>
-          </Block>
-
-          <Block marginBlockStart="space020" marginBlockEnd="space050">
-            <FormInput
-              name="country"
-              rules={{
-                required: 'Required field',
-              }}
-            >
-              <FormInputLabel>Country</FormInputLabel>
-              <FormInputSelect>
-                <SelectOption value="BG">Bulgaria</SelectOption>
-                <SelectOption value="UK">United Kingdom</SelectOption>
-                <SelectOption value="NL">The Netherlands</SelectOption>
-                <SelectOption value="DE">Germany</SelectOption>
-              </FormInputSelect>
-              <FormInputAssistiveText>
-                Enter your last name
-              </FormInputAssistiveText>
-            </FormInput>
-          </Block>
-        </Fieldset>
-
-        <Fieldset legend="Login information">
-          <Block marginBlockStart="space020" marginBlockEnd="space050">
-            <FormInput
-              name="email"
-              rules={{
-                required: 'Required field',
-                pattern: /^\S+@\S+$/i,
-              }}
-            >
-              <FormInputLabel>Email</FormInputLabel>
-              <FormInputTextField type="email" />
-              <FormInputAssistiveText>Enter your email</FormInputAssistiveText>
-            </FormInput>
-          </Block>
-
-          <Block marginBlockStart="space020" marginBlockEnd="space050">
-            <FormInput
-              name="password"
-              rules={{
-                required: 'Required field',
-                minLength: {
-                  value: 8,
-                  message: 'Name must be at least 8 characters long',
-                },
-              }}
-            >
-              <FormInputLabel>Password</FormInputLabel>
-              <FormInputTextField
-                type={showPassword ? 'text' : 'password'}
-                endEnhancer={
-                  <ShowPasswordButton
-                    onClick={() => toggleShowPassword(!showPassword)}
-                    isVisible={showPassword}
-                  />
-                }
-              />
-              <FormInputAssistiveText>
-                Enter your password
-              </FormInputAssistiveText>
-            </FormInput>
-          </Block>
-        </Fieldset>
-
-        <Fieldset legend="Gender">
-          <Block marginBlockStart="space020" marginBlockEnd="space050">
-            <FormInput
-              name="sex"
-              rules={{
-                required: 'Required field',
-              }}
-            >
-              <RadioGroup>
-                <FormInputRadioButton
-                  id="gender-male"
-                  value="male"
-                  label="Male"
-                  overrides={{marginBlockEnd: 'space020'}}
-                />
-                <FormInputRadioButton
-                  id="gender-female"
-                  value="female"
-                  label="Female"
-                  overrides={{marginBlockEnd: 'space020'}}
-                />
-                <FormInputRadioButton
-                  id="gender-3rd"
-                  value="3rd"
-                  label="Third Gender"
-                  overrides={{marginBlockEnd: 'space020'}}
-                />
-                <FormInputRadioButton
-                  id="gender-no-say"
-                  value="noSay"
-                  label="Not to say"
-                  overrides={{marginBlockEnd: 'space020'}}
-                />
-              </RadioGroup>
-              <FormInputAssistiveText>Make your choice</FormInputAssistiveText>
-            </FormInput>
-          </Block>
-        </Fieldset>
-
-        <Fieldset legend="Interests">
-          <Block marginBlockStart="space020" marginBlockEnd="space050">
-            <FormInput
-              name="interests"
-              rules={{
-                required: 'Required field',
-                validate: (value: string[]) => {
-                  if (value.length < 3) {
-                    return 'Select at least 3 interests';
-                  }
-
-                  return true;
-                },
-              }}
-            >
-              <FormInputCheckbox
-                id="interests-politics"
-                value="politics"
-                label="Politics"
-                overrides={{marginBlockEnd: 'space020'}}
-              />
-              <FormInputCheckbox
-                id="interests-business"
-                value="Business"
-                label="Business"
-                overrides={{marginBlockEnd: 'space020'}}
-              />
-              <FormInputCheckbox
-                id="interests-society"
-                value="Society"
-                label="Society"
-                overrides={{marginBlockEnd: 'space020'}}
-              />
-              <FormInputCheckbox
-                id="interests-technology"
-                value="Technology"
-                label="Technology"
-                overrides={{marginBlockEnd: 'space020'}}
-              />
-              <FormInputCheckbox
-                id="interests-sport"
-                value="Sport"
-                label="Sport"
-                overrides={{marginBlockEnd: 'space020'}}
-              />
-              <FormInputCheckbox
-                id="interests-science"
-                value="Science"
-                label="Science"
-                overrides={{marginBlockEnd: 'space020'}}
-              />
-              <FormInputAssistiveText>Make your choice</FormInputAssistiveText>
-            </FormInput>
-          </Block>
-        </Fieldset>
-
-        <Button type="submit">Register</Button>
-      </Form>
-    </FormContainer>
-  );
-};
-StoryFormComplete.storyName = 'Form complete';
 
 export default {
   title: 'Components/Form input',
